+++ src/Selah.Domain/Data/Models/Transactions/Commands/TransactionCreateResponse.cs
@@ -8,7 +8,7 @@ namespace Selah.Domain.Data.Models.Transactions.Commands
 {
     public class TransactionCreateResponse
     {
-        public long TransactionId { get; set; }
+        public int TransactionId { get; set; }
 
         public DateTime TransactionDate { get; set; }
