+++ src/Selah.Domain/Data/Models/Transactions/TransactionLineItemCreate.cs
@@ -4,9 +4,9 @@ namespace Selah.Domain.Data.Models.Transactions
 {
     public class TransactionLineItemCreate
     {
-        public long TransactionId { get; set; }
+        public int TransactionId { get; set; }
 
-        public long? TransactionCategoryId { get; set; }
+        public int? TransactionCategoryId { get; set; }
 
         public decimal ItemizedAmount { get; set; }
