+++ src/Selah.Domain/Data/Models/Transactions/UserTransactionResponse.cs
@@ -6,7 +6,7 @@ namespace Selah.Domain.Data.Models.Transactions
   public class UserTransactionResponse
   {
     public string TransactionDate { get; set; }
-    public long Records { get; set; }
+    public int Records { get; set; }
     public List<UserTransactionVM> Transactions { get; set; }
     
     public decimal TotalTransactionAmount { get; set; }
